simplify

+++ News/src/main/java/com/hzpd/services/InitService.java
@@ -10,7 +10,7 @@ import com.hzpd.hflt.BuildConfig;
 import com.hzpd.modle.UserBean;
 import com.hzpd.modle.db.UserLog;
 import com.hzpd.modle.db.UserLogDao;
-import com.hzpd.ui.App;
+import com.hzpd.ui.ConfigBean;
 import com.hzpd.url.InterfaceJsonfile;
 import com.hzpd.url.OkHttpClientManager;
 import com.hzpd.utils.DBHelper;
@@ -18,13 +18,17 @@ import com.hzpd.utils.FjsonUtil;
 import com.hzpd.utils.Log;
 import com.hzpd.utils.RequestParamsUtils;
 import com.hzpd.utils.SPUtil;
+import com.hzpd.utils.SharePreferecesUtils;
 import com.news.update.UpdateUtils;
 import com.news.update.Utils;
+import com.squareup.okhttp.OkHttpClient;
 import com.squareup.okhttp.Request;
+import com.squareup.okhttp.Response;
 
 import java.io.File;
 import java.io.IOException;
 import java.security.NoSuchAlgorithmException;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -32,12 +36,12 @@ public class InitService extends IntentService {
 
     public static final String InitAction = "initService";
     public static final String UserLogAction = "user.log.action";
-    public static final String SHARE_KEY_AD = "key_ad";
+    public static final String SHARE_KEY_AD_CONFIG = "key_config";
+    public static final String SHARE_CONFIG_MODIFIED_TIME = "key_config_modified_time";
 
-    private String rootPath;
-    String siteid = null;
-    String CACHE_url = null;
-    String mAdPic_url = null;
+    public static long SEND_TIME = 1000 * 60 * 2;
+    public static long send_user_log_time = System.currentTimeMillis();
+
     private Object tag;
 
     public InitService() {
@@ -47,16 +51,15 @@ public class InitService extends IntentService {
     @Override
     public void onCreate() {
         super.onCreate();
-        rootPath = App.getInstance().getJsonFileCacheRootDir();
-        siteid = InterfaceJsonfile.SITEID;
-        CACHE_url = InterfaceJsonfile.CACHE;
-        mAdPic_url = InterfaceJsonfile.mAdPic;
+        SEND_TIME = ConfigBean.getInstance().send_log_time;
         tag = OkHttpClientManager.getTag();
     }
 
     @Override
     public void onDestroy() {
         super.onDestroy();
+        OkHttpClientManager.cancel(tag);
+        Log.e("test", "News: " + "onDestroy");
     }
 
     @Override
@@ -66,6 +69,7 @@ public class InitService extends IntentService {
             if (InitService.InitAction.equals(intent.getAction())) {
                 debugTest();
                 //TODO 获取服务器配置  AD_KEY
+                getConfig();
             } else if (InitService.UserLogAction.equals(intent.getAction())) {
                 Log.e("test", "send use log");
                 sendUserLog();
@@ -73,8 +77,35 @@ public class InitService extends IntentService {
         }
     }
 
+    private void getConfig() {
+        try {
+            if (!Utils.isNetworkConnected(this)) {
+                return;
+            }
+            Map<String, String> params = new HashMap<>();
+            SPUtil.addParams(params);
+            Request request = new Request.Builder().url(InterfaceJsonfile.AD_CONFIG).head().build();
+            Response response = new OkHttpClient().newCall(request).execute();
+            Log.e("test", "News: " + response.headers());
+            String data = OkHttpClientManager.post(InterfaceJsonfile.AD_CONFIG, params);
+            if (!TextUtils.isEmpty(data)) {
+                com.alibaba.fastjson.JSONObject json = FjsonUtil.parseObject(data);
+                if (json.getString("code").equals("200")) {
+                    Log.e("test", "News: " + json);
+                    SharePreferecesUtils.setParam(this, SHARE_KEY_AD_CONFIG, data);
+                    ConfigBean.update();
+                }
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+
     private void sendUserLog() {
         if (BuildConfig.DEBUG || !Utils.isNetworkConnected(this)) {
+            return;
+        }
+        if (System.currentTimeMillis() - send_user_log_time < SEND_TIME) {
             return;
         }
 
@@ -86,7 +117,7 @@ public class InitService extends IntentService {
         final UserLogDao dbUtils = DBHelper.getInstance(getApplicationContext()).getLog();
         try {
             final List<UserLog> logs = dbUtils.loadAll();
-            if (logs == null || logs.isEmpty() || logs.size() < 20) {
+            if (logs == null || logs.isEmpty()) {
                 return;
             }
             String json = FjsonUtil.toJsonString(logs);
@@ -95,7 +126,8 @@ public class InitService extends IntentService {
             params.put("json", json);
             SPUtil.addParams(params);
             String str = OkHttpClientManager.post(InterfaceJsonfile.USER_LOG, params);
-            if (!str.isEmpty()) {
+            if (!str.isEmpty() && str.contains("200")) {
+                send_user_log_time = System.currentTimeMillis();
                 dbUtils.deleteAll();
             }
         } catch (Exception e) {
@@ -135,8 +167,6 @@ public class InitService extends IntentService {
             };
             File file = Environment.getExternalStorageDirectory();
             file = new File(file, "news.test.apk");
-
-            //OkHttpClientManager.download(url, file, true, callBack);
         }
     }
 }
